make private traits in public api, sealed and public
